Truncate IPM steps to ensure sign condition. Use state IMPLIED_EQ only for dependent eq constraints.
Iterate::Update() guarantees that moving xl, xu, zl and zu keeps the
variable positive if it has a barrier term in the IPM. The safeguard is
necessary because for infeasible models the step directions can be huge
so that in finite precision arithmetic the step size from the IPM ratio
test can make a variable zero.

Iterate::StateDetail::IMPLIED_EQ is intented for slack variables to
dependent equality constraints. (It has never been used for anything
else.) Iterate::make_implied_eq() now sets zl[j]=zu[j]=0, so that this
need not be done in starting_basis.cc:PostprocessDependencies(). In
constrast, make_implied_lb() and make_implied_ub(), which are used for
fixing variables during the IPM solve, zl[j] and zu[j] are kept at their
current value.