Perform model scaling after dualization.
Change the order of preprocessing steps for building the computational
model from the user input: first dualize (if appropriate), then scale.
Doing scaling at the end is preferable for two reasons:

(1) We can store the original user model rather than the scaled user
    model. This simplifies code for computing infeasibilities to the
    original user model.

(2) It is more natural to scale after dualization because then the
    scaling method can treat right-hand side and objective differently.
    Although we do not exploit this yet.

As a byproduct, negating variables with infinite lower bound but finite
upper bound is now only done for dualization. Not required otherwise.