Fix postsolve bug.
This bug appeared when dualizing models that have at least one variable
with finite upper bound but infinite lower bound. Such variables are
flipped when building the dual problem. In postsolve we first have to
build the solution to the user model with flipped variables. For that we
must not look at the bounds of the original user model.